requirements.txt
pyserial
psutil
PyQt5
sqlmodel